+++ NEWS.md
@@ -12,6 +12,8 @@
   - DouglasPeuckerLineSimplifier, avoid crash with Point input and NaN tolerance (GH-1078, Dan Baston)
   - GEOSLineSubstring, avoid crash with NaN length fraction (GH-1077, Dan Baston)
   - MinimumClearance, avoid crash on NaN inputs (GH-1079, Dan Baston)
+  - Fix LineString->getPoint(n) for M geometries (GH-1191, @hsieyuan)
+
 
 ## Changes in 3.12.2
 2024-06-05